Removed users code, added stats
